Move `mapError` function to `internal.cmd` to be shared.
